ud036 : Lesson 1 : Secret Message : process file names
I watched the lesson as far as listing the directory, and decided to research the method I would use to do it if I was on my own.

This stage covers the deduction of the new file name.  The next stage would be to modify the files, or generate an output page that displays them in order.

Signed-off-by: James Arlow <james.arlow@gmail.com>